+++ mlrun/datastore/datastore.py
@@ -191,7 +191,7 @@ class StoreManager:
                 raise ValueError(f"no such store ({endpoint})")
 
         store_key = f"{schema}://{endpoint}"
-        if not secrets and not mlrun.is_running_as_api():
+        if not secrets and not mlrun.utils.is_running_as_api():
             if store_key in self._stores.keys():
                 return self._stores[store_key], subpath
 
@@ -201,6 +201,6 @@ class StoreManager:
         store = schema_to_store(schema)(
             self, schema, store_key, parsed_url.netloc, secrets=secrets
         )
-        if not secrets and not mlrun.is_running_as_api():
+        if not secrets and not mlrun.utils.is_running_as_api():
             self._stores[store_key] = store
         return store, subpath
